Tools to download posts from Tumblr
This tool creates JSON files based on a single search of a blog's
posts. It will create one file per post found. Notes on the post
include tags on any reblogs, which are looked up on the separate read
only API to avoid authenticated rate limits.

There is also a heuristic to put the notes into threaded conversations
which atm is not returned by the API. Replies are linked where the
heuristic thinks that one comment was a reply to another. This also
happens for reblogs – if it looks like a comment could only be replying
to a reblog's tags.