Correctly understand the parent post even for reblogs that don't add any text
The `trail` property only contains reblogs that add non-tag content. So
if we interrogate the trail we miss the fact that a reblog that has
reblogged another reblog. The `parent_post_url` is always correct
though, so we can use that. The format should probably be assumed to
be opaque, but we have nothing from the open API to compare it with
so we can extract the post ID for now.