Make sure to collect all notes
The `with_notes` API feature only returns a limited number of notes,
and they can't be paged through. The `notes` endpoint can return all
of the notes and we already know how to page through them. So once we
have the post (which is now quicker because notes aren't loaded), go
and get the notes explicitly to make sure we capture them all.