Use the `notes` endpoint more strategically
The various `modes` on the `notes` endpoint return different fields –
notably, a note that has dded text does not need to have its
reblogged from ID looked up, as it's included as part of one of the
modes. So now only do extra lookups if we actually need the info that
we would get from that lookup.