Correctly match replies to replies using timestamps
The old method just ran through the replies in order, and tries to
pair them up. This basically assumes that they are served in order but
it's clear now that they aren't. So now we order them by timestamp
explicitly, and just search backwards for the previous post. This works
for long comment chains, but may break down if one commentator writes
multiple comments – all replies written after the last post will hang
off the last post, even if they were replying to the first.

Resolves #1.