Get missing links for the masterpost
The masterpost isn't always updated, so we shouldn't rely on it to get
all post links, but we can use it to fill in any gaps we've found as
anything in the masterpost should definitely be archived.

We do it this way around because its way more efficient in terms of
requests to do a search (20 posts per request), rather than having
a list of IDs and iterating through them (1 post per request).