Cluster archiving around posting times
Analysis of the notes shows that the majority of notes are sent in the
first 48 hours after a page is posted. The schedule is thus updated to
focus on Mondays and now additionally Tuesdays, and then drop back to
a 12 hour schedule for other days.

We also no longer update from a push, as code testing is done. If we
need to test this again in the future, we can just turn pushing back
on.